added separete api notifications
