+++ api/index.js
@@ -63,6 +63,7 @@ const register = (server, options, next) => {
   server.register(require('./inspection_support'))
   server.register(require('./search'))
   server.register(require('./v2/cases'))
+  server.register(require('./notifications'))
 
   server.ext('onPreResponse', preResponse)
   server.ext('onRequest', onRequest)
