+++ api/notifications/package.json
@@ -1,0 +1,4 @@
+{
+    "name": "notifications",
+    "versopm": "1.0.0"
+}
